refactor: schedule을 booking 내부로 이동
