+++ src/main/java/roomescape/booking/reservation/ReservationRepository.java
@@ -1,9 +1,9 @@
 package roomescape.booking.reservation;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import roomescape.booking.schedule.Schedule;
 import roomescape.member.Member;
 import roomescape.reservationtime.ReservationTime;
-import roomescape.schedule.Schedule;
 import roomescape.theme.Theme;
 
 import java.time.LocalDate;
