+++ src/main/java/roomescape/booking/schedule/ScheduleController.java
@@ -1,4 +1,4 @@
-package roomescape.schedule;
+package roomescape.booking.schedule;
 
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.schedule.dto.ScheduleRequest;
-import roomescape.schedule.dto.ScheduleResponse;
+import roomescape.booking.schedule.dto.ScheduleRequest;
+import roomescape.booking.schedule.dto.ScheduleResponse;
 
 @RestController
 @RequestMapping("/schedules")
