+++ src/main/java/roomescape/booking/waiting/WaitingRepository.java
@@ -2,8 +2,8 @@ package roomescape.booking.waiting;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import roomescape.booking.schedule.Schedule;
 import roomescape.member.Member;
-import roomescape.schedule.Schedule;
 
 import java.util.List;
