+++ src/test/java/roomescape/reservation/reservation/ReservationServiceTest.java
@@ -14,14 +14,14 @@ import roomescape.booking.reservation.dto.AdminFilterReservationRequest;
 import roomescape.booking.reservation.dto.AdminReservationRequest;
 import roomescape.booking.reservation.dto.ReservationRequest;
 import roomescape.booking.reservation.dto.ReservationResponse;
+import roomescape.booking.schedule.Schedule;
+import roomescape.booking.schedule.ScheduleRepository;
 import roomescape.exception.custom.reason.reservation.ReservationConflictException;
 import roomescape.exception.custom.reason.reservation.ReservationNotExistsMemberException;
 import roomescape.member.Member;
 import roomescape.member.MemberRepository;
 import roomescape.member.MemberRole;
 import roomescape.reservationtime.ReservationTime;
-import roomescape.schedule.Schedule;
-import roomescape.schedule.ScheduleRepository;
 import roomescape.theme.Theme;
 
 import java.time.LocalDate;
