+++ src/test/java/roomescape/waiting/WaitingServiceTest.java
@@ -8,6 +8,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import roomescape.auth.dto.LoginMember;
 import roomescape.booking.reservation.Reservation;
 import roomescape.booking.reservation.ReservationRepository;
+import roomescape.booking.schedule.Schedule;
+import roomescape.booking.schedule.ScheduleRepository;
 import roomescape.booking.waiting.Waiting;
 import roomescape.booking.waiting.WaitingRepository;
 import roomescape.booking.waiting.WaitingService;
@@ -19,8 +21,6 @@ import roomescape.member.Member;
 import roomescape.member.MemberRepository;
 import roomescape.member.MemberRole;
 import roomescape.reservationtime.ReservationTime;
-import roomescape.schedule.Schedule;
-import roomescape.schedule.ScheduleRepository;
 import roomescape.theme.Theme;
 
 import java.time.LocalDate;
